Convert deprecated numpy data types to built-in types
Numpy deprecated and removed the `numpy.float`, `numpy.int`,
`numpy.complex` and `numpy.long` data types. For the first three
they were just aliases for the built-in types so they can safely
converted to the built-in types in this commit. `numpy.long` is
changed the appropriate scalar type; `numpy.longlong`.

In all the files touched, duplicate numpy imports are removed.
Typically there was both `import numpy as np` and `import numpy
as n`. The former is preferred and used in all files touched by
this commit.